move delegations specific functionality to own file
